// src/generated/UsageHistoryPageDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UsageHistoryCursorDto } from "./UsageHistoryCursorDto";
import type { UsageHistoryRowDto } from "./UsageHistoryRowDto";

export type UsageHistoryPageDto = { rows: Array<UsageHistoryRowDto>, nextCursor: UsageHistoryCursorDto | null, totalRows: number, };